Measure the runtime

+++ python_async_function/2-measure_runtime.py
@@ -14,4 +14,4 @@ def measure_time(n: int, max_delay: int) -> float:
     asyncio.run(wait_n(n, max_delay))
     end = time.perf_counter()
     total_time = end - start
-    return total_time / n
+    return float(total_time / n)
